test_project/requirements.txt
#
# Primary requirements
#
Django>=2.2.3,<3.2.13
djangorestframework>=3.9.0,<3.13.1
djangorestframework-simplejwt==5.2.0
django_extensions==3.1.5            # Needed to run shell_plus

#
# Secondary requirements
#
django-phonenumber-field==6.1.0     # Needed by django-contact
phonenumbers==8.12.49               # Needed by django-contact
